requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
pydantic==2.11.2
pydantic_core==2.33.1
setuptools==78.1.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
torch==2.6.0
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
python-jose==3.4.0
requests==2.31.0
python-dotenv==1.0.1
cryptography==42.0.5